+++ module-application/src/main/java/com/devtoon/jtoon/payment/presentation/PaymentController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.devtoon.jtoon.iamport.application.IamportService;
 import com.devtoon.jtoon.payment.application.PaymentInfoService;
 import com.devtoon.jtoon.payment.request.PaymentInfoDto;
 import com.siot.IamportRestClient.exception.IamportResponseException;
@@ -18,14 +19,19 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/payments")
-public class PaymentInfoController {
+public class PaymentController {
 
+	private final IamportService iamportService;
 	private final PaymentInfoService paymentInfoService;
 
-	@PostMapping
-	public IamportResponse<Payment> verifyIamport(@RequestBody PaymentInfoDto paymentInfoDto)
+	@PostMapping("/validation")
+	public IamportResponse<Payment> validateIamport(@RequestBody PaymentInfoDto paymentInfoDto)
 		throws IamportResponseException, IOException {
-		IamportResponse<Payment> iamportResponse = paymentInfoService.paymentLookup(paymentInfoDto.impUid());
-		return paymentInfoService.verifyIamport(iamportResponse, paymentInfoDto);
+		return iamportService.validateIamport(paymentInfoDto.impUid(), paymentInfoDto.amount());
+	}
+
+	@PostMapping
+	public void createPaymentInfo(@RequestBody PaymentInfoDto paymentInfoDto) {
+		paymentInfoService.createPaymentInfo(paymentInfoDto);
 	}
 }
