+++ module-application/src/main/resources/application.yml
@@ -1,3 +1,3 @@
 spring:
   profiles:
-    include: s3, smtp
+    include: s3, smtp, iamport
