+++ module-application/src/main/resources/templates/payments.html
@@ -26,7 +26,6 @@
             }, function (rsp) {
                 // 결제 성공 시
                 if (rsp.success) {
-                    console.log(rsp);
                     let data = {
                         impUid: rsp.imp_uid,
                         merchantUid: rsp.merchant_uid,
@@ -42,13 +41,31 @@
                     // 서버로 해당 정보로 검증 및 주문서 생성 요청
                     $.ajax({
                         type: "POST",
-                        url: 'api/v1/payments',
+                        url: 'api/v1/payments/validation',
                         data: JSON.stringify(data),
                         contentType: "application/json",
                         dataType: "json",
+
                         success: function (result) {
-                            alert("결제 검증 완료");
-                        }, error: function (result) {
+                            $.ajax({
+                                type: "POST",
+                                url: 'api/v1/payments',
+                                data: JSON.stringify(data),
+                                contentType: "application/json",
+                                dataType: "json",
+
+                                success: function (result) {
+                                    alert("결제 검증 완료");
+                                },
+
+                                error: function (result) {
+                                    console.log(jqXHR.status + ", " + textStatus + ", " + errorThrown);
+                                    alert("3. 결제 실패 : " + jqXHR.status + ", " + errorThrown);
+                                }
+                            });
+                        },
+
+                        error: function (result) {
                             console.log(jqXHR.status + ", " + textStatus + ", " + errorThrown);
                             alert("2. 결제 실패 : " + jqXHR.status + ", " + errorThrown);
                         }
