+++ module-domain-iamport/build.gradle
@@ -3,15 +3,9 @@ repositories {
 }
 
 dependencies {
-    // Web
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
     // I-AM-PORT
     implementation 'com.github.iamport:iamport-rest-client-java:0.2.23'
-
+    
     // Bean Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
-
-    // Test
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
