+++ module-domain/build.gradle
@@ -1,27 +1,16 @@
-repositories {
-    maven { url 'https://jitpack.io' }
-}
-
 dependencies {
     // JPA
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation project(path: ':module-domain-smtp')
-
-    // I-AM-PORT
-    implementation 'com.github.iamport:iamport-rest-client-java:0.2.23'
-
-    // Bean Validation
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    
-    // Test
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
     // H2
-    testImplementation 'com.h2database:h2'
+    implementation 'com.h2database:h2'
 
     // Bean Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
     // Security
     implementation 'org.springframework.boot:spring-boot-starter-security'
+
+    // Test
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
