--- module-domain/src/test/resources/application-test.yml
@@ -1,17 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:~/jtoon;MODE=MYSQL;DB_CLOSE_ON_EXIT=FALSE;
-    username: sa
-    password:
-    driver-class-name: org.h2.Driver
-
-  jpa:
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-
-    database: h2
-
-logging.level:
-    org.hibernate.SQL: debug
