+++ settings.gradle
@@ -5,3 +5,5 @@ include 'module-domain-s3'
 include 'module-domain-smtp'
 include 'module-internal'
 include 'module-core'
+include 'module-domain-iamport'
+
